separete server and auth
