+++ backend/internal/config/auth.go
@@ -1,0 +1,8 @@
+package config
+
+import "time"
+
+type AuthToken struct {
+	SecretKey string        `yaml:"secret_key"`
+	Expires   time.Duration `yaml:"expires"`
+}
