+++ backend/internal/config/server.go
@@ -1,0 +1,6 @@
+package config
+
+type Server struct {
+	Host string `yaml:"host"`
+	Port string `yaml:"port"`
+}
